Factor out TimeRangeFilter
